MOD00-DesafiosLG/desafio-5-automação-de-monitoramento-do-dólar-com-atualização-diária-em-excel-usando-dados-do-site-do-banco-central/monitora_dolar_bot/requirements.txt
beautifulsoup4==4.12.3
botcity-framework-base==0.4.4
botcity-framework-web==0.9.0
botcity-maestro-sdk==0.5.1
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
et_xmlfile==2.0.0
fonttools==4.54.1
idna==3.10
importlib_metadata==8.5.0
kiwisolver==1.4.7
matplotlib==3.9.2
msedge-selenium-tools==3.141.3
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
selenium==3.141.0
six==1.16.0
soupsieve==2.6
tzdata==2024.2
urllib3==1.26.20
webdriver-manager==4.0.2
zipp==3.20.2
